# common-platform-accounting-gl-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-accounting-gl
  description: General Ledger Accounting Management
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/general-ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/journal-batches/{batchId}/entries/{entryId}/lines/{lineId}:
    get:
      tags:
        - Journal Lines
      summary: Get a Journal Line by ID
      operationId: getById_2
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: lineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JournalLineDTO'
    put:
      tags:
        - Journal Lines
      summary: Update a Journal Line
      operationId: update_2
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: lineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalLineDTO'
        required: true
      responses:
        '200':
          description: Line updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JournalLineDTO'
    delete:
      tags:
        - Journal Lines
      summary: Delete a Journal Line
      operationId: delete_2
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: lineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Line deleted
...
  /api/v1/journal-batches/{batchId}/entries/{entryId}/lines:
    post:
      tags:
        - Journal Lines
      summary: Create a new Journal Line in a specific Entry
      operationId: create_2
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalLineDTO'
        required: true
      responses:
        '201':
          description: Journal Line created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JournalLineDTO'
...
  /api/v1/journal-batches/{batchId}/entries/{entryId}/lines/:
    get:
      tags:
        - Journal Lines
      summary: Search/Filter Journal Lines
      operationId: search_2
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestJournalLineDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseJournalLineDTO'
...
components:
  schemas:
...
    JournalLineDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entryId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        amount:
          type: number
        drCr:
          type: string
          enum:
            - DR
            - CR
        memo:
          type: string
        fxCurrency:
          type: string
        fxRate:
          type: number
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
    FilterRequestJournalLineDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/JournalLineDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponseJournalLineDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/JournalLineDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...